+++ andes/models/__init__.py
@@ -30,7 +30,7 @@ file_classes = OrderedDict([
     ('measurement', ['BusFreq', 'BusROCOF', 'PMU']),
     ('dcbase', ['Node', 'Ground', 'R', 'L', 'C', 'RCp', 'RCs', 'RLs', 'RLCs', 'RLCp']),
     ('vsc', ['VSCShunt']),
-    ('renewable', ['REGCA1', 'REECA1', 'REECA1I',
+    ('renewable', ['REGCA1', 'REECA1', 'REECA1E',
                    'REPCA1', 'WTDTA1', 'WTDS', 'WTARA1', 'WTPTA1', 'WTTQA1', 'WTARV1',
                    'REGCVSG', 'REGCVSG2']),
     ('distributed', ['PVD1', 'ESD1']),